# edge-control-plane/cmd/api/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: edgeCloud Control Plane API
  version: 1.0.0
  description: |
    HTTP API for the edgeCloud managed WebAssembly edge computing platform.
    Supports tenant and worker authentication. Tenants authenticate via
    Bearer API key. Workers authenticate via Bearer JWT (HMAC-SHA256).
servers:
  - url: https://api.edgecloud.dev
    description: Production
  - url: http://localhost:8080
    description: Local development
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: Tenant API key (Bearer token). Obtain via POST /api/v1/keys.
    WorkerJWTAuth:
      type: http
      scheme: bearer
      description: Worker JWT (HMAC-SHA256). Issued to workers by the control plane.
    InternalTokenAuth:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: |
        Shared-secret header for trusted service-to-service callers
        (peer control planes pulling artifacts, the edge-ingress
        fetching traffic splits). Token value must match the
        receiving control plane's `internal_token` config. Reject
        requests to `http://` peer URLs — TLS is required to protect
        the token in transit.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code.
              example: NOT_FOUND
            message:
              type: string
              description: Human-readable error detail. Always "internal error" for 5xx.
              example: deployment not found

    # -------------------------------------------------------------------------
    # Health
    # -------------------------------------------------------------------------
...
    BootstrapRequest:
      type: object
      required: [name, key_name]
      properties:
        name:
          type: string
          description: Unique tenant name.
          example: acme
        plan:
          type: string
          default: free
          description: Subscription plan.
          example: free
        key_name:
          type: string
          description: Name for the first API key.
          example: default
    BootstrapResponse:
      type: object
      properties:
        tenant_id:
          type: string
          description: The new tenant's ID.
          example: t_abc123
        api_key:
          type: string
          description: |
            The raw API key. Shown only once — store it securely.
            Use Authorization: Bearer <api_key> for all subsequent requests.
          example: ek_...
    CreateTenantRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: acme
        plan:
          type: string
          example: free
    Tenant:
      type: object
      properties:
        id:
          type: string
          example: t_abc123
        name:
          type: string
          example: acme
        plan:
          type: string
          example: free
        allowlisted_destinations:
          type: array
          items:
            type: string
          example: ["api.stripe.com"]
        created_at:
          type: string
          format: date-time
          example: "2026-01-02T03:04:05Z"
    UpdateTenantRequest:
      type: object
      properties:
        name:
          type: string
          example: acme-inc
        plan:
          type: string
          example: pro
        allowlisted_destinations:
          type: array
          items:
            type: string
          example: ["api.stripe.com", "api.github.com"]

    # -------------------------------------------------------------------------
    # API Keys
    # -------------------------------------------------------------------------
...
  responses:
    BadRequest:
      description: Bad request — invalid or missing parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: BAD_REQUEST
              message: name is required
    Unauthorized:
      description: Unauthorized — missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: invalid api key
    Forbidden:
      description: Forbidden — insufficient permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: FORBIDDEN
              message: forbidden
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: deployment not found
...
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INTERNAL_ERROR
              message: internal error
...
paths:
...
  /api/v1/tenants:
    post:
      operationId: bootstrapTenant
      summary: Self-signup — create tenant and first API key
      description: |
        Creates a new tenant and atomically issues the first API key.
        No authentication required. The returned raw API key is shown only once.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapRequest'
            example:
              name: acme
              key_name: default
      responses:
        '201':
          description: Tenant and API key created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/admin/tenants:
    get:
      operationId: listTenants
      summary: List all tenants (admin only)
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: List of all tenants.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tenant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: createTenant
      summary: Create a tenant (admin only)
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/admin/tenants/{tenantID}:
    get:
      operationId: getTenant
      summary: Get a tenant by ID (admin only)
      security:
        - ApiKeyAuth: []
      parameters:
        - name: tenantID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tenant details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      operationId: updateTenant
      summary: Update a tenant (admin only)
      security:
        - ApiKeyAuth: []
      parameters:
        - name: tenantID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenantRequest'
      responses:
        '200':
          description: Tenant updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      operationId: deleteTenant
      summary: Delete a tenant and all its data (admin only)
      security:
        - ApiKeyAuth: []
      parameters:
        - name: tenantID
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Tenant deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  # -------------------------------------------------------------------------
  # Admin: Apps
  # -------------------------------------------------------------------------
...